+++ crates/uv/src/commands/mod.rs
@@ -35,7 +35,6 @@ pub(crate) use python::dir::dir as python_dir;
 pub(crate) use python::find::find as python_find;
 pub(crate) use python::install::install as python_install;
 pub(crate) use python::list::list as python_list;
-pub(crate) use python::patch::patch as python_patch;
 pub(crate) use python::pin::pin as python_pin;
 pub(crate) use python::uninstall::uninstall as python_uninstall;
 #[cfg(feature = "self-update")]
