+++ docs/reference/cli.md
@@ -3790,8 +3790,6 @@ uv python [OPTIONS] <COMMAND>
 </dd>
 <dt><a href="#uv-python-pin"><code>uv python pin</code></a></dt><dd><p>Pin to a specific Python version</p>
 </dd>
-<dt><a href="#uv-python-patch"><code>uv python patch</code></a></dt><dd><p>Update the patches of distributions specified in .python-version(s) files to the latest available versions</p>
-</dd>
 <dt><a href="#uv-python-dir"><code>uv python dir</code></a></dt><dd><p>Show the uv Python installation directory</p>
 </dd>
 <dt><a href="#uv-python-uninstall"><code>uv python uninstall</code></a></dt><dd><p>Uninstall Python versions</p>
@@ -4293,108 +4291,7 @@ uv python pin [OPTIONS] [REQUEST]
 
 <p>This option is usually not safe to use when committing the <code>.python-version</code> file to version control.</p>
 
-</dd><dt><code>--verbose</code>, <code>-v</code></dt><dd><p>Use verbose output.</p>
-
-<p>You can configure fine-grained logging using the <code>RUST_LOG</code> environment variable. (&lt;https://docs.rs/tracing-subscriber/latest/tracing_subscriber/filter/struct.EnvFilter.html#directives&gt;)</p>
-
-</dd><dt><code>--version</code>, <code>-V</code></dt><dd><p>Display the uv version</p>
-
-</dd></dl>
-
-### uv python patch
-
-Update the patches of distributions specified in .python-version(s) files to the latest available versions
-
-<h3 class="cli-reference">Usage</h3>
-
-```
-uv python patch [OPTIONS]
-```
-
-<h3 class="cli-reference">Options</h3>
-
-<dl class="cli-reference"><dt><code>--cache-dir</code> <i>cache-dir</i></dt><dd><p>Path to the cache directory.</p>
-
-<p>Defaults to <code>$HOME/Library/Caches/uv</code> on macOS, <code>$XDG_CACHE_HOME/uv</code> or <code>$HOME/.cache/uv</code> on Linux, and <code>%LOCALAPPDATA%\uv\cache</code> on Windows.</p>
-
-<p>May also be set with the <code>UV_CACHE_DIR</code> environment variable.</p>
-</dd><dt><code>--color</code> <i>color-choice</i></dt><dd><p>Control colors in output</p>
-
-<p>[default: auto]</p>
-<p>Possible values:</p>
-
-<ul>
-<li><code>auto</code>:  Enables colored output only when the output is going to a terminal or TTY with support</li>
-
-<li><code>always</code>:  Enables colored output regardless of the detected environment</li>
-
-<li><code>never</code>:  Disables colored output</li>
-</ul>
-</dd><dt><code>--config-file</code> <i>config-file</i></dt><dd><p>The path to a <code>uv.toml</code> file to use for configuration.</p>
-
-<p>While uv configuration can be included in a <code>pyproject.toml</code> file, it is not allowed in this context.</p>
-
-<p>May also be set with the <code>UV_CONFIG_FILE</code> environment variable.</p>
-</dd><dt><code>--directory</code> <i>directory</i></dt><dd><p>Change to the given directory prior to running the command.</p>
-
-<p>Relative paths are resolved with the given directory as the base.</p>
-
-<p>See <code>--project</code> to only change the project root directory.</p>
-
-</dd><dt><code>--help</code>, <code>-h</code></dt><dd><p>Display the concise help for this command</p>
-
-</dd><dt><code>--native-tls</code></dt><dd><p>Whether to load TLS certificates from the platform&#8217;s native certificate store.</p>
-
-<p>By default, uv loads certificates from the bundled <code>webpki-roots</code> crate. The <code>webpki-roots</code> are a reliable set of trust roots from Mozilla, and including them in uv improves portability and performance (especially on macOS).</p>
-
-<p>However, in some cases, you may want to use the platform&#8217;s native certificate store, especially if you&#8217;re relying on a corporate trust root (e.g., for a mandatory proxy) that&#8217;s included in your system&#8217;s certificate store.</p>
-
-<p>May also be set with the <code>UV_NATIVE_TLS</code> environment variable.</p>
-</dd><dt><code>--no-cache</code>, <code>-n</code></dt><dd><p>Avoid reading from or writing to the cache, instead using a temporary directory for the duration of the operation</p>
-
-<p>May also be set with the <code>UV_NO_CACHE</code> environment variable.</p>
-</dd><dt><code>--no-config</code></dt><dd><p>Avoid discovering configuration files (<code>pyproject.toml</code>, <code>uv.toml</code>).</p>
-
-<p>Normally, configuration files are discovered in the current directory, parent directories, or user configuration directories.</p>
-
-<p>May also be set with the <code>UV_NO_CONFIG</code> environment variable.</p>
-</dd><dt><code>--no-progress</code></dt><dd><p>Hide all progress outputs.</p>
-
-<p>For example, spinners or progress bars.</p>
-
-</dd><dt><code>--no-python-downloads</code></dt><dd><p>Disable automatic downloads of Python.</p>
-
-</dd><dt><code>--offline</code></dt><dd><p>Disable network access.</p>
-
-<p>When disabled, uv will only use locally cached data and locally available files.</p>
-
-</dd><dt><code>--project</code> <i>project</i></dt><dd><p>Run the command within the given project directory.</p>
-
-<p>All <code>pyproject.toml</code>, <code>uv.toml</code>, and <code>.python-version</code> files will be discovered by walking up the directory tree from the project root, as will the project&#8217;s virtual environment (<code>.venv</code>).</p>
-
-<p>Other command-line arguments (such as relative paths) will be resolved relative to the current working directory.</p>
-
-<p>See <code>--directory</code> to change the working directory entirely.</p>
-
-<p>This setting has no effect when used in the <code>uv pip</code> interface.</p>
-
-</dd><dt><code>--python-preference</code> <i>python-preference</i></dt><dd><p>Whether to prefer uv-managed or system Python installations.</p>
-
-<p>By default, uv prefers using Python versions it manages. However, it will use system Python installations if a uv-managed Python is not installed. This option allows prioritizing or ignoring system Python installations.</p>
-
-<p>May also be set with the <code>UV_PYTHON_PREFERENCE</code> environment variable.</p>
-<p>Possible values:</p>
-
-<ul>
-<li><code>only-managed</code>:  Only use managed Python installations; never use system Python installations</li>
-
-<li><code>managed</code>:  Prefer managed Python installations over system Python installations</li>
-
-<li><code>system</code>:  Prefer system Python installations over managed Python installations</li>
-
-<li><code>only-system</code>:  Only use system Python installations; never use managed Python installations</li>
-</ul>
-</dd><dt><code>--quiet</code>, <code>-q</code></dt><dd><p>Do not print any output</p>
+</dd><dt><code>--upgrade</code></dt><dd><p>Upgrade the patches of distributions specified in .python-version(s) files to the latest available versions</p>
 
 </dd><dt><code>--verbose</code>, <code>-v</code></dt><dd><p>Use verbose output.</p>
